Write all unit tests for quote and ticker
Unit tests for services include connection to database
and are written for edge cases too.

+++ src/quotes/entities/quote.entity.ts
@@ -6,7 +6,7 @@ export class QuoteEntity {
   @PrimaryColumn({ type: 'int' })
   timestamp: number;
 
-  @Column({ type: 'numeric' })
+  @Column({ type: 'float' })
   price: number;
 
   @PrimaryColumn({ type: 'varchar', length: 6 })
